+++ src/_services/stats.js
@@ -5,6 +5,8 @@ import internet from 'internet-available';
 import got from 'got';
 import logger from 'winston';
 import prettyBytes from 'pretty-bytes';
+import { asyncWriteFile } from '../_utils/files';
+import {resolve} from 'path';
 
 let intervalID;
 
@@ -29,8 +31,9 @@ export async function sendPayload() {
 		}
 		const payload = await buildPayload();
 		logger.info(`[Stats] Sending payload (${prettyBytes(JSON.stringify(payload).length)})`);
-		logger.debug(`[Stats] Payload being sent : ${JSON.stringify(payload,null,2)}`);
+		logger.debug('[Stats] Payload data saved locally to logs/statsPayload.json');
 		const conf = getConfig();
+		asyncWriteFile(resolve(conf.appPath, 'logs/statsPayload.json'), JSON.stringify(payload, null, 2), 'utf-8');
 		try {
 			await got(`https://${conf.OnlineHost}/api/stats`,{
 				json: true,