+++ src/_services/welcome.js
@@ -1,7 +1,7 @@
 import {getConfig} from '../_utils/config';
 import {editUser} from '../_services/user';
 import randomstring from 'randomstring';
-import opn from 'opn';
+import open from 'open';
 
 function generateAdminPassword() {
 	// Resets admin's password when appFirstRun is set to 1.
@@ -22,11 +22,11 @@ export async function welcomeToYoukousoKaraokeMugen(port) {
 	const conf = getConfig();
 	if (+conf.appFirstRun === 1) {
 		const adminPassword = generateAdminPassword();
-		if (!conf.optNoBrowser && !conf.isDemo && !conf.isTest) opn(`http://localhost:${port}/welcome?admpwd=${adminPassword}`);
+		if (!conf.optNoBrowser && !conf.isDemo && !conf.isTest) open(`http://localhost:${port}/welcome?admpwd=${adminPassword}`);
 		console.log(`\nAdmin password is : ${adminPassword}\nPlease keep it in a safe place, it will not be displayed ever again.\nTo reset admin password, set appFirstRun to 1 in config.ini\n`);
 	} else {
 		if (!conf.optNoBrowser && !conf.isDemo && !conf.isTest) {
-			opn(`http://localhost:${port}/welcome`);
+			open(`http://localhost:${port}/welcome`);
 		}
 	}
 }