+++ src/_webapp/ressources/js/karaokemugen.js
@@ -89,8 +89,8 @@ var settingsNotUpdated;
 			'TAGCAT_FAMI':['TAG_ANIME','TAG_REAL','TAG_VIDEOGAME'],
 			'TAGCAT_SUPP':['TAG_3DS','TAG_DREAMCAST','TAG_DS','TAG_GAMECUBE','TAG_PC','TAG_PS2','TAG_PS3','TAG_PS4','TAG_PSP','TAG_PSV','TAG_PSX','TAG_SATURN','TAG_SEGACD','TAG_SWITCH','TAG_WII','TAG_WII','TAG_XBOX360'],
 			'TAGCAT_CLAS':['TAG_IDOL','TAG_MAGICALGIRL','TAG_MECHA','TAG_SHOUJO','TAG_SHOUNEN','TAG_YAOI','TAG_YURI'],
-			'TAGCAT_ORIG':['TAG_MOBAGE','TAG_MOVIE','TAG_ONA','TAG_OVA','TAG_TOKU','TAG_TVSHOW','TAG_VN','TAG_VOCALOID'],
-			'TAGCAT_TYPE':['TAG_DUO','TAG_HARDMODE','TAG_HUMOR','TAG_LONG','TAG_PARODY','TAG_R18','TAG_REMIX','TAG_SPECIAL','TAG_SPOIL'],
+			'TAGCAT_ORIG':['TAG_MOBAGE','TAG_DRAMA','TAG_MOVIE','TAG_ONA','TAG_OVA','TAG_TOKU','TAG_TVSHOW','TAG_VN','TAG_VOCALOID'],
+			'TAGCAT_TYPE':['TAG_DUO','TAG_HARDMODE','TAG_HUMOR','TAG_LONG','TAG_PARODY','TAG_R18','TAG_COVER','TAG_REMIX','TAG_SPECIAL','TAG_SPOIL'],
 		}
 		flattenedTagsGroups = [].concat.apply([], Object.values(tagsGroups));
 		// Once page is loaded
@@ -395,6 +395,47 @@ var settingsNotUpdated;
 				}).done(function () {
 					DEBUG && console.log('Playlist ' + idPlaylist + ' emptied');
 				});
+			} else if (name === 'addRandomKaras') {
+
+				displayModal('prompt', i18n.__('CL_ADD_RANDOM_TITLE'),'', function(nbOfRandoms){
+					$.ajax({
+						url: url,
+						data : { random : nbOfRandoms},
+						type: 'GET'
+					}).done(function (randomKaras) {
+						// console.log(randomKaras);
+						if(randomKaras.content.length > 0) {
+
+							let textContent = randomKaras.content.map(e => buildKaraTitle(e)).join('<br/><br/>');
+
+							displayModal('confirm', i18n.__('CL_CONGRATS'), i18n.__('CL_ABOUT_TO_ADD', '<br/><br/>' + textContent), function(){
+								var karaList = randomKaras.content.map(function(a) {
+									return a.kid;
+								}).join();
+
+
+				            	var urlPost = getPlData(idPlaylistTo).url;
+
+								$.ajax({
+									url: urlPost,
+									type: 'POST',
+									data: { kid : karaList },
+									complete: function() {
+										$('#modalBox').modal('hide');
+										$('body').removeClass('modal-open');
+										$('.modal-backdrop').remove();
+									}
+								}).done(function () {
+									DEBUG && console.log(karaList + ' added to playlist ' + idPlaylistTo);
+
+								});
+							},'');
+						} else {
+							console.log('Error : server could not pick any random song');
+						}
+					});
+				},'');
+
 			}
 		});
 
@@ -585,22 +626,24 @@ var settingsNotUpdated;
 		$('.getLucky').on('click', function () {
 			var filter = $('#searchPlaylist' + 1).val();
 
-			$.ajax({ url: 'public/karas/random?filter=' + filter }).done(function (data) {
-				var chosenOne = data;
-				$.ajax({ url: 'public/karas/' + chosenOne }).done(function (data) {
-					data = data[0];
-					displayModal('confirm', i18n.__('CL_CONGRATS'), i18n.__('CL_ABOUT_TO_ADD', buildKaraTitle(data)), function(){
-						$.ajax({
-							url: 'public/karas/' + chosenOne,
-							type: 'POST',
-							data: { requestedby : logInfos.username }
-						}).done(function () {
-							playlistContentUpdating.done( function() {
-								scrollToKara(2, chosenOne);
+			$.ajax({ url: 'public/karas?filter=' + filter, data : { random : 1 } }).done(function (data) {
+				if(data && data.content && data.content[0]) {
+					var chosenOne = data.content[0].kid;
+					$.ajax({ url: 'public/karas/' + chosenOne }).done(function (data) {
+						data = data[0];
+						displayModal('confirm', i18n.__('CL_CONGRATS'), i18n.__('CL_ABOUT_TO_ADD', buildKaraTitle(data)), function(){
+							$.ajax({
+								url: 'public/karas/' + chosenOne,
+								type: 'POST',
+								data: { requestedby : logInfos.username }
+							}).done(function () {
+								playlistContentUpdating.done( function() {
+									scrollToKara(2, chosenOne);
+								});
 							});
-						});
-					},'lucky');
-				});
+						},'lucky');
+					});
+				}
 			});
 		});
 		$('.favorites').on('click', function() {
@@ -1102,7 +1145,7 @@ var settingsNotUpdated;
 	addKaraHtml = '<button title="' + i18n.__('TOOLTIP_ADDKARA')
                 + (scope == 'admin' ? ' - ' + i18n.__('TOOLTIP_ADDKARA_ADMIN') : '')
                 + '" name="addKara" class="btn btn-sm btn-action"></button>';
-                
+
 	deleteKaraHtml = '<button title="' + i18n.__('TOOLTIP_DELETEKARA') + '" name="deleteKara" class="btn btn-sm btn-action"></button>';
 	deleteCriteriaHtml = '<button title="' + i18n.__('TOOLTIP_DELETECRITERIA') + '" name="deleteCriteria" class="btn btn-action deleteCriteria"></button>';
 	transferKaraHtml = '<button title="' + i18n.__('TOOLTIP_TRANSFERKARA') + '" name="transferKara" class="btn btn-sm btn-action"></button>';
@@ -1326,7 +1369,7 @@ var settingsNotUpdated;
 					'tag' : 't'
 				}[searchCriteria]
 				: '';
-            
+
 			urlFiltre += '&searchType=' + searchType + '&searchValue=' + (searchCriteria && searchValue ? searchCriteria + ':' + searchValue : '');
 		}
 
@@ -1384,7 +1427,7 @@ var settingsNotUpdated;
 								if (kara.flag_upvoted) {
 									likeKara = likeKaraHtml.replace('likeKara', 'likeKara currentLike');
 								}
-                                
+
 								// TODO add fav button next to info for public pc interface
 								htmlContent += '<li class="list-group-item" ' + karaDataAttributes + '>'
 								//	+ 	(scope == 'public' && isTouchScreen ? '<slide></slide>' : '')
@@ -1503,7 +1546,7 @@ var settingsNotUpdated;
 								});
 								var tagText = tagsFiltered.length === 1 && data[k].type > 0  && data[k].type < 100 ?  tagsFiltered[0].name_i18n : data[k].value;
 								var textContent = data[k].type == 1001 ? buildKaraTitle(data[k].value[0]) : tagText;
-    
+
 								blacklistCriteriasHtml.find('li[type="' + data[k].type + '"]').after(
 									'<li class="list-group-item liTag" blcriteria_id="' + data[k].blcriteria_id + '"> '
                                 +	'<div class="actionDiv">' + html + '</div>'
@@ -1511,7 +1554,7 @@ var settingsNotUpdated;
                                 +	'<div class="contentDiv">' + textContent + '</div>'
                                 +	'</li>');
 							})
-						
+
 						}
 					}
 					//htmlContent = blacklistCriteriasHtml.html();
@@ -1690,7 +1733,7 @@ var settingsNotUpdated;
 			var searchOptionListHtml = '<option value="-1" default data-playlist_id="-1"></option>';
 			searchOptionListHtml += '<option value="-6" data-playlist_id="-6"></option>';
 			searchOptionListHtml += '<option value="-5" data-playlist_id="-5" data-flag_favorites="true"></option>';
-            
+
 			// building the options
 			var optionListHtml = '';
 			$.each(playlistList, function (key, value) {
@@ -1869,7 +1912,7 @@ var settingsNotUpdated;
 
 
 				if ( data.currentlyPlaying === null) {
-                    
+
 					$('#karaInfo').attr('idKara', data.currentlyPlaying);
 					$('#karaInfo').attr('length', -1);
 					$('#karaInfo > span').text( i18n.__('KARA_PAUSED_WAITING') );
@@ -1943,7 +1986,7 @@ var settingsNotUpdated;
 		if(data.languages && isMulti) {
 			data.languages = [isMulti];
 		}
-        
+
 		var titleText = 'fillerTitle';
 
 		var limit = isSmall ? 35 : 50;
@@ -2264,7 +2307,7 @@ var settingsNotUpdated;
 
 		return $option;
 	};
-    
+
 	formatTagsPlaylist = function (playlist) {
 		if (!playlist.id) return playlist.text;
 
@@ -2287,10 +2330,10 @@ var settingsNotUpdated;
 		var searchVal1 = '', searchVal2 = '';
 		if(locSearchPlaylist1 && locSearchPlaylist1 != 'undefined') searchVal1 = locSearchPlaylist1;
 		if(locSearchPlaylist2 && locSearchPlaylist2 != 'undefined') searchVal2 = locSearchPlaylist2;
-                
+
 		$('#searchPlaylist1').val(searchVal1);
 		$('#searchPlaylist2').val(searchVal2);
-    
+
 
 		setupAjax();
 
@@ -2570,7 +2613,7 @@ var settingsNotUpdated;
 		var deferred = $.Deferred();
 		var url = 'auth/login';
 		var data = { username: username, password: password};
-        
+
 		if(!username) {
 			url = 'auth/login/guest';
 			data = { fingerprint : password };
@@ -2705,18 +2748,18 @@ var settingsNotUpdated;
 	manageOnlineUsersUI = function(data) {
 		$('[name="modalLoginServ"]').val(data['OnlineUsers'] ? data['OnlineHost'] : '');
 		DEBUG && console.log(logInfos);
-		
+
 		settingsUpdating.done(function () {
-            if(!settings.OnlineUsers || !data.OnlineUsers || logInfos.onlineToken || logInfos.role == 'guest') {
-                $('.profileConvert').hide();
-            } else {
-                $('.profileConvert').show();
-            }
-            if(logInfos.onlineToken) {
-                $('.profileDelete').show();
-            } else {
-                $('.profileDelete').hide();
-            }
+			if(!settings.OnlineUsers || !data.OnlineUsers || logInfos.onlineToken || logInfos.role == 'guest') {
+				$('.profileConvert').hide();
+			} else {
+				$('.profileConvert').show();
+			}
+			if(logInfos.onlineToken) {
+				$('.profileDelete').show();
+			} else {
+				$('.profileDelete').hide();
+			}
 			if(!data.OnlineUsers && (Object.keys(settings).length == 0 || settings.OnlineUsers) && logInfos.username.includes('@')) {
 				setTimeout(function() {
 					displayMessage('warning',i18n.__('LOG_OFFLINE.TITLE') + '<br/>', i18n.__('LOG_OFFLINE.MESSAGE'), 8000);